Make command mode transfers work the same as other transfers.
The local and UDP transfer functions both return the status of the
transfer itself.  The command mode returns CVME_IO if the transfer
failed, but if the module failed it also would return that error code.
This modification stuffs the result code into the buffer as expected and
returns 0 if the transfer succeeded.