Output the out of scope fact in qmail and vmailmgr CVMs
When a qmail lookup is done on a domain that does not exist in
virtualdomains, no determination can be made about users in that domain,
so the result is "out of scope".  Similarly, vmailmgr treats any failed
qmail lookup as "out of scope", since it can't then do its virtual user
lookup.

This fact is also reported on error in cvm-testclient.

Also bumped the version to 0.85, due to the significance of this change.